+++ src/components/_share/JinniProvider/JinniProvider.tsx
@@ -5,10 +5,8 @@ import JinniContext, {
 import { getJinniBreakPointValue } from '@/utils/breakpoint';
 import { getJinniColorValue } from '@/utils/color';
 import { getJinniTypographyValue } from '@/utils/typography';
-import {
-  getJinniBoxShadowValue,
-  getJinniWhiteOverlayValue
-} from '@/utils/elevation';
+import { getJinniWhiteOverlayValue } from '@/utils/overlay';
+import { getJinniBoxShadowValue } from '@/utils/boxShadow';
 import { BREAKPOINTS } from '@/constants/breakpoint';
 import { COLOR_THEME, COLOR_PALETTE } from '@/constants/color';
 import { TYPOGRAPHY } from '@/constants/typography';
@@ -61,14 +59,17 @@ const JinniProvider = ({ children }: JinniProviderProps) => {
   const functions = {
     getElevation: (elevationLevel: ElevationLevelType) => {
       const { boxShadow, whiteOverlay } = designSystem;
-      if (themeMode === 'dark')
-        return {
-          boxShadow: boxShadow[elevationLevel],
-          backgroundImage: whiteOverlay[elevationLevel]
-        };
-      return {
-        boxShadow: boxShadow[elevationLevel]
-      };
+      switch (themeMode) {
+        case 'light':
+          return {
+            boxShadow: boxShadow[elevationLevel]
+          };
+        case 'dark':
+          return {
+            boxShadow: boxShadow[elevationLevel],
+            backgroundImage: whiteOverlay[elevationLevel]
+          };
+      }
     }
   };
