+++ src/utils/boxShadow.ts
@@ -13,21 +13,3 @@ export const getJinniBoxShadowValue = (elevationLevel: ElevationLevelType) => {
 
   return boxShadowValue;
 };
-
-export const getJinniWhiteOverlayValue = (
-  elevationLevel: ElevationLevelType
-) => {
-  const rootEl = document.querySelector(':root');
-  if (!rootEl) throw Error('root element를 가져오지 못함');
-
-  const rootStyles = window.getComputedStyle(rootEl);
-  const whiteOverlayValue = rootStyles.getPropertyValue(
-    `--jinni-white-overlay-${elevationLevel}`
-  );
-  if (!whiteOverlayValue)
-    throw Error(
-      `'${elevationLevel}' level의 white-overlay는 존재하지 않습니다`
-    );
-
-  return whiteOverlayValue;
-};
